rename savePartyDetailsCurrent.js to savePartyDetails to be consistent with function name; move into src directory
